python/requirements.txt
mcpf-python>=1.0.0-alpha
